requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
openai
chromadb
pydantic-settings
pytest
httpx